8/11/2022

+++ src/StaticThing.java
@@ -13,7 +13,7 @@ public class StaticThing {
         return desert;
     }
 
-    public StaticThing(String fileName, int x){
+    public StaticThing(String fileName, double x){
         Image fichierDesert = new Image(fileName);  //ouverture du fichier png
         desert = new ImageView(fichierDesert);      //chargement de l'image
         desert.setViewport(new Rectangle2D(0, 50,800,300)); //découpage de l'image aux dimensions souhaitées
